Suppress deadcode warnings from static analysis tools
It's intentional that trunnel will generate some dead checks: I
deliberately wrote it to _always_ check how many bytes are
remaining, even if the check might be dead.  The alternative would
be to make the code more complex in order to teach it how to
sometimes omit checks... but such complexity would run the risk of
accidentally omitting a necessary check.

So replace the "if (remaining < foo) goto bar;" checks with a macro,
and use an #ifdef to define the macro such that the static tool
(probably) won't conclude that check can never be true.